+++ firmware/hw_layer/ports/stm32/stm32f1/hw_ports.mk
@@ -1,4 +1,5 @@
 HW_LAYER_EMS += $(PROJECT_DIR)/hw_layer/ports/stm32/flash.c
 
 HW_LAYER_EMS_CPP += $(PROJECT_DIR)/hw_layer/ports/stm32/stm32f1/mpu_util.cpp \
+	$(PROJECT_DIR)/hw_layer/ports/stm32/stm32_pins.cpp \
 	$(PROJECT_DIR)/hw_layer/ports/stm32/stm32_common.cpp
